r.volume: change int type to prevent overflow, see #3004 (merge from trunk, r68376)
git-svn-id: https://svn.osgeo.org/grass/grass/branches/releasebranch_7_0@68378 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ raster/r.volume/local_proto.h
@@ -1,3 +1,3 @@
 #define SEP "-----------------------------------------------------------------------------"
 
-int centroids(int, int *, int *, int, int);
+int centroids(int, unsigned long*, unsigned long*, int, int);
